Separated event handling classes from GameManager
