+++ src/main/java/net/SimpleSurvival/GameManager.java
@@ -298,74 +298,3 @@ public class GameManager implements Listener {
     }
 }
 
-class GameStarter extends BukkitRunnable {
-
-    private final SimpleSurvival plugin;
-    private GameManager currentGame;
-    private int countdown;
-
-    public GameStarter(SimpleSurvival plugin, GameManager currentGame, int countdown) {
-        this.plugin = plugin;
-        this.currentGame = currentGame;
-        this.countdown = countdown;
-    }
-
-    @Override
-    public void run() {
-        for (Player p: Bukkit.getWorld(currentGame.getWorldUUID()).getPlayers()) {
-            p.sendMessage("The game will begin in " + countdown + " seconds!");
-        }
-        countdown--;
-        if (countdown <= 0) {
-            for (Player p: Bukkit.getWorld(currentGame.getWorldUUID()).getPlayers()) {
-                p.sendMessage("The game has begun!");
-            }
-            currentGame.setState(GameManager.GameState.RUNNING);
-            this.cancel();
-        }
-    }
-}
-
-class GameEnder extends BukkitRunnable {
-
-    private final SimpleSurvival plugin;
-    private GameManager currentGame;
-    private int countdown;
-
-    public GameEnder(SimpleSurvival plugin, GameManager currentGame, int countdown) {
-        this.plugin = plugin;
-        this.currentGame = currentGame;
-        this.countdown = countdown;
-
-        for (Player p : Bukkit.getWorld(this.currentGame.getWorldUUID()).getPlayers()) {
-            p.sendMessage("The world will close in " + countdown + " seconds!");
-        }
-    }
-
-    @Override
-    public void run() {
-        countdown--;
-        if (countdown <= 0) {
-            for (Player p : Bukkit.getWorld(currentGame.getWorldUUID()).getPlayers()) {
-                p.sendMessage("World Closing...");
-            }
-            for (Player p : Bukkit.getWorld(currentGame.getWorldUUID()).getPlayers()) {
-                p.setGameMode(GameMode.SURVIVAL);
-                p.setAllowFlight(false);
-                p.getInventory().clear();
-                for(Player player : Bukkit.getWorld(currentGame.getWorldUUID()).getPlayers()) {
-                    p.showPlayer(player);
-                }
-            }
-
-            for(Player p : Bukkit.getWorld(currentGame.getWorldUUID()).getPlayers()) {
-                p.teleport(Bukkit.getServer().getWorlds().get(0).getSpawnLocation());
-            }
-
-            plugin.worldManager.destroyWorld(currentGame.getWorldUUID());
-            currentGame.unregisterListeners();
-            this.cancel();
-        }
-
-    }
-}
